requirements.txt
# Flask Backend Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
Werkzeug==3.0.1

# Google OAuth
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Email Support
Flask-Mail==0.9.1

# File Upload
Pillow==10.1.0

# Production Server
gunicorn==21.2.0
